Migrate config files to .mjs/.ts, update Tailwind and add next-themes dependency
