+++ jest.config.js
@@ -1,9 +1,9 @@
-const nextJest = require("next/jest")
+const nextJest = require("next/jest");
 
 const createJestConfig = nextJest({
-  // Provide the path to your Next.js app to load next.config.js and .env files in your test environment
+  // Provide the path to your Next.js app to load next.config.mjs and .env files in your test environment
   dir: "./",
-})
+});
 
 // Add any custom config to be passed to Jest
 const customJestConfig = {
@@ -18,7 +18,7 @@ const customJestConfig = {
     "<rootDir>/.next/",
     "<rootDir>/__tests__/utils/", // Ignore the utils directory
   ],
-}
+};
 
 // createJestConfig is exported this way to ensure that next/jest can load the Next.js config which is async
-module.exports = createJestConfig(customJestConfig)
+module.exports = createJestConfig(customJestConfig);
