--- next.config.js
@@ -1,20 +1,0 @@
-const { i18n } = require("./next-i18next.config")
-
-const nextConfig = {
-  typescript: {
-    ignoreBuildErrors: true,
-  },
-  reactStrictMode: true,
-  i18n,
-  images: {
-    remotePatterns: [
-      {
-        protocol: "https",
-        hostname: "**.public.blob.vercel-storage.com",
-      },
-    ],
-  },
-}
-
-module.exports = nextConfig
-
