+++ package.json
@@ -21,6 +21,7 @@
     "lucide-react": "^0.279.0",
     "next": "^14.0.3",
     "next-i18next": "^14.0.3",
+    "next-themes": "^0.4.6",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-i18next": "^13.2.2",
